#92 Add booking session start/end and activity set queries
Introduces endpoints and handlers for starting and ending booking sessions, including session start/end time tracking. Adds query and specification to fetch activity sets by session activity ID. Updates DTOs and mapping profiles to support new fields and type changes for practice/rest times. Refactors related command handlers and configuration for session status.

+++ FitBridge_Application/Dtos/ActivitySets/ActivitySetUpdateRequestDto.cs
@@ -8,6 +8,6 @@ public class ActivitySetUpdateRequestDto
     public double? WeightLifted { get; set; }
     public int? NumOfReps { get; set; }
     public bool IsCompleted { get; set; }
-    public int? PracticeTime { get; set; }
-    public int? RestTime { get; set; }
+    public double? PracticeTime { get; set; }
+    public double? RestTime { get; set; }
 }
